+++ nethsm/client/apis/path_to_api.py
@@ -1,11 +1,15 @@
 import typing
 import typing_extensions
 
+from nethsm.client.apis.paths.cluster_join import ClusterJoin
+from nethsm.client.apis.paths.cluster_members import ClusterMembers
+from nethsm.client.apis.paths.cluster_members_member_id import ClusterMembersMemberID
 from nethsm.client.apis.paths.config_backup_passphrase import ConfigBackupPassphrase
 from nethsm.client.apis.paths.config_logging import ConfigLogging
 from nethsm.client.apis.paths.config_network import ConfigNetwork
 from nethsm.client.apis.paths.config_time import ConfigTime
 from nethsm.client.apis.paths.config_tls_cert_pem import ConfigTlsCertPem
+from nethsm.client.apis.paths.config_tls_cluster_ca_pem import ConfigTlsClusterCaPem
 from nethsm.client.apis.paths.config_tls_csr_pem import ConfigTlsCsrPem
 from nethsm.client.apis.paths.config_tls_generate import ConfigTlsGenerate
 from nethsm.client.apis.paths.config_tls_public_pem import ConfigTlsPublicPem
@@ -52,11 +56,15 @@ from nethsm.client.apis.paths.users_user_id_tags_tag import UsersUserIDTagsTag
 PathToApi = typing.TypedDict(
     'PathToApi',
     {
+    "/cluster/join": typing.Type[ClusterJoin],
+    "/cluster/members": typing.Type[ClusterMembers],
+    "/cluster/members/{MemberID}": typing.Type[ClusterMembersMemberID],
     "/config/backup-passphrase": typing.Type[ConfigBackupPassphrase],
     "/config/logging": typing.Type[ConfigLogging],
     "/config/network": typing.Type[ConfigNetwork],
     "/config/time": typing.Type[ConfigTime],
     "/config/tls/cert.pem": typing.Type[ConfigTlsCertPem],
+    "/config/tls/cluster-ca.pem": typing.Type[ConfigTlsClusterCaPem],
     "/config/tls/csr.pem": typing.Type[ConfigTlsCsrPem],
     "/config/tls/generate": typing.Type[ConfigTlsGenerate],
     "/config/tls/public.pem": typing.Type[ConfigTlsPublicPem],
@@ -104,11 +112,15 @@ PathToApi = typing.TypedDict(
 
 path_to_api = PathToApi(
     {
+    "/cluster/join": ClusterJoin,
+    "/cluster/members": ClusterMembers,
+    "/cluster/members/{MemberID}": ClusterMembersMemberID,
     "/config/backup-passphrase": ConfigBackupPassphrase,
     "/config/logging": ConfigLogging,
     "/config/network": ConfigNetwork,
     "/config/time": ConfigTime,
     "/config/tls/cert.pem": ConfigTlsCertPem,
+    "/config/tls/cluster-ca.pem": ConfigTlsClusterCaPem,
     "/config/tls/csr.pem": ConfigTlsCsrPem,
     "/config/tls/generate": ConfigTlsGenerate,
     "/config/tls/public.pem": ConfigTlsPublicPem,