+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 
 ## Unreleased
 
--
+### Features
+
+- Add support for clustering:
+  - Add `ClusterMember`, `InitialClusterMember` and `ClusterJoinData` classes
+  - Add `add_cluster_member`, `list_cluster_members`, `set_cluster_member_urls`, `remove_cluster_member`, `join_cluster`, `get_cluster_ca_certificate`, `set_cluster_ca_certificate` functions to `NetHSM`
 
 [All Changes](https://github.com/Nitrokey/nethsm-sdk-py/compare/v2.0.1...HEAD)